Move ReferenceSymbols class to ReferenceSymbols.cs
